Move crypto price and farcaster frame spec outside of component files
