+++ components/common/CharmEditor/specRegistry.ts
@@ -8,11 +8,11 @@ import { spec as bookmarkSpec } from './components/bookmark/bookmarkSpec';
 import * as bulletList from './components/bulletList';
 import * as callout from './components/callout/callout';
 import * as columnLayout from './components/columnLayout/columnLayout.schema';
-import { cryptoPriceSpec } from './components/CryptoPrice';
+import { spec as cryptoPriceSpec } from './components/cryptoPrice/cryptoPriceSpec';
 import * as disclosure from './components/disclosure';
 import * as doc from './components/doc/doc';
 import * as emoji from './components/emojiSuggest/emojiSuggest.specs';
-import { farcasterFrameSpec } from './components/FarcasterFrame';
+import { spec as farcasterFrameSpec } from './components/farcasterFrame/farcasterFrameSpec';
 import { deletion, formatChange, insertion } from './components/fiduswriter/schema/common/track';
 import { specs as fileSpecs } from './components/file/file.specs';
 import * as heading from './components/heading';
